backend/requirements.txt
openai-whisper
google-generativeai
python-dotenv
Flask>=3.0.0